code_generator/pyproject.toml
[project]
name = "code-generator"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "kourin", email = "kourin.code@gmail.com" }
]
dependencies = [
    "guidance~=0.0.63",
    "python-dotenv~=1.0.0",
    "langchain~=0.0.27",
    "openai~=0.27.8",
    "backoff~=2.2.1",
    "google-search-results~=2.4.2",
    "Flask~=2.3.2",
    "CORS~=1.0.1",
    "flask-cors~=3.0.10",
    "tinydb~=4.8.0",
    "celery~=5.3.1",
    "Redis~=4.5.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
